pipelines/pyproject.toml
[project]
name = "pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "imagecodecs>=2023.3.16",
    "mercantile>=1.2.1",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=10.4.0",
    "pmtiles>=3.4.1",
    "python-ulid>=1.1.0",
    "rioxarray>=0.13.4",
    "scipy>=1.10.1",
    "shapely>=2.0.7",
]
